chore: Refactoring and organizing tests. #275
